Break up spin-test main file
